Unit tests for BrowserStack app uploads
